@@ -8,7 +8,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "aap-eda"
-version = "0.1.0"
+version = "0.2.0"
 description = ""
 authors = ["Red Hat, Inc. <info@ansible.com>"]
 readme = "README.md"
@@ -25,27 +25,29 @@ all = ["psycopg2"]
 dev = ["psycopg2-binary"]
 
 [tool.poetry.dependencies]
-python = ">=3.9,<4"
+python = ">=3.11,<3.12"
 django = ">=4.2,<4.3"
-djangorestframework = "*"
-dynaconf = ">=3"
-drf-spectacular = "*"
-channels = { version = "*", extras = ["daphne"] }
-psycopg2 = { version = "*", optional = true }
+djangorestframework = "3.14.*"
+djangorestframework_simplejwt = "5.3.*"
+dynaconf = ">=3.1.12,<3.3"
+drf-spectacular = ">=0.26.5,<0.27"
+channels = { version = "4.0.*", extras = ["daphne"] }
+psycopg2 = { version = "2.9.*", optional = true }
 psycopg2-binary = { version = "*", optional = true }
-django-rq = "*"
-rq = "*"
-django-filter = "*"
-pydantic = "*"
-cryptography = "*"
-kubernetes = "*"
-podman = ">=4.5"
+django-rq = ">=2.5,<2.9"
+rq = ">=1.10.1,<1.14"
+django-filter = ">23.2,<24"
+pydantic = ">=1.8.1,<1.11"
+cryptography = ">=38,<42"
+kubernetes = "26.1.*"
+podman = "4.9.*"
 rq-scheduler = "^0.10"
-# Experimental LDAP Integration https://issues.redhat.com/browse/AAP-16938
-# We are temporarily updating it to use latest from devel branch to keep consistency accross
-# other AAP components
-django-ansible-base = { git = "https://github.com/ansible/django-ansible-base.git" }
-jinja2 = "*"
+django-ansible-base = { git = "https://github.com/ansible/django-ansible-base.git", tag = "2024.2.12", extras = [
+    "channel_auth",
+] }
+jinja2 = ">=3.1.3,<3.2"
+django-split-settings = "^1.2.0"
+pexpect = "^4.9.0"
 
 [tool.poetry.group.test.dependencies]
 pytest = "*"
@@ -109,6 +111,7 @@ extend-ignore = [
     "S101",   # Asserts allowed in tests
     "ARG",    # Fixtures are not always used explicitly
     "SLF001", # Call private methods in tests
+    "D",      # Docstrings are not required in tests
 ]
 
 [tool.ruff.flake8-tidy-imports]
