@@ -14,7 +14,7 @@
 
 from rest_framework import serializers
 
-from aap_eda.api.serializers.credential import CredentialRefSerializer
+from aap_eda.api.serializers.eda_credential import EdaCredentialRefSerializer
 from aap_eda.core import models
 
 
@@ -30,7 +30,7 @@ class DecisionEnvironmentSerializer(serializers.ModelSerializer):
             "name",
             "description",
             "image_url",
-            "credential_id",
+            "eda_credential_id",
             *read_only_fields,
         ]
 
@@ -38,7 +38,9 @@ class DecisionEnvironmentSerializer(serializers.ModelSerializer):
 class DecisionEnvironmentCreateSerializer(serializers.ModelSerializer):
     """Serializer for creating the DecisionEnvironment."""
 
-    credential_id = serializers.IntegerField(required=False, allow_null=True)
+    eda_credential_id = serializers.IntegerField(
+        required=False, allow_null=True
+    )
 
     class Meta:
         model = models.DecisionEnvironment
@@ -46,14 +48,16 @@ class DecisionEnvironmentCreateSerializer(serializers.ModelSerializer):
             "name",
             "description",
             "image_url",
-            "credential_id",
+            "eda_credential_id",
         ]
 
 
 class DecisionEnvironmentReadSerializer(serializers.ModelSerializer):
     """Serializer for reading the DecisionEnvironment with embedded objects."""
 
-    credential = CredentialRefSerializer(required=False, allow_null=True)
+    eda_credential = EdaCredentialRefSerializer(
+        required=False, allow_null=True
+    )
 
     class Meta:
         model = models.DecisionEnvironment()
@@ -62,16 +66,18 @@ class DecisionEnvironmentReadSerializer(serializers.ModelSerializer):
             "name",
             "description",
             "image_url",
-            "credential",
+            "eda_credential",
             "created_at",
             "modified_at",
         ]
         read_only_fields = ["id", "created_at", "modified_at"]
 
     def to_representation(self, decision_environment):
-        credential = (
-            CredentialRefSerializer(decision_environment["credential"]).data
-            if decision_environment["credential"]
+        eda_credential = (
+            EdaCredentialRefSerializer(
+                decision_environment["eda_credential"]
+            ).data
+            if decision_environment["eda_credential"]
             else None
         )
         return {
@@ -79,7 +85,7 @@ class DecisionEnvironmentReadSerializer(serializers.ModelSerializer):
             "name": decision_environment["name"],
             "description": decision_environment["description"],
             "image_url": decision_environment["image_url"],
-            "credential": credential,
+            "eda_credential": eda_credential,
             "created_at": decision_environment["created_at"],
             "modified_at": decision_environment["modified_at"],
         }
