@@ -15,6 +15,7 @@
 from drf_spectacular.utils import extend_schema_field
 from rest_framework import serializers
 
+from aap_eda.api.serializers.fields.yaml import YAMLSerializerField
 from aap_eda.core import models
 
 
@@ -47,129 +48,6 @@ class RulebookRefSerializer(serializers.ModelSerializer):
         model = models.Rulebook
         fields = ["id", "name", "description"]
         read_only_fields = ["id"]
-
-
-class RulesetSerializer(serializers.ModelSerializer):
-    name = serializers.CharField(
-        required=True,
-        help_text="Name of the ruleset",
-    )
-
-    sources = serializers.JSONField(
-        required=True,
-        help_text="The contained sources in the ruleset",
-    )
-
-    class Meta:
-        model = models.Ruleset
-        fields = "__all__"
-        read_only_fields = ["id", "created_at", "modified_at"]
-
-
-class RulesetOutSerializer(serializers.Serializer):
-    id = serializers.IntegerField(
-        required=True,
-        help_text="ID of the ruleset",
-    )
-
-    name = serializers.CharField(
-        required=True,
-        help_text="Name of the ruleset",
-    )
-
-    rule_count = serializers.IntegerField(
-        required=True,
-        help_text="Number of rules the ruleset contains",
-    )
-
-    source_types = serializers.ListField(
-        child=serializers.CharField(),
-        required=True,
-        help_text="List of source types",
-    )
-
-    fired_stats = serializers.ListField(
-        child=serializers.JSONField(),
-        required=True,
-        help_text="List of stats",
-    )
-
-    created_at = serializers.DateTimeField(
-        required=True,
-        help_text="The created_at timestamp of the ruleset",
-    )
-
-    modified_at = serializers.DateTimeField(
-        required=True,
-        help_text="The modified_at timestamp of the ruleset",
-    )
-
-
-class RuleSerializer(serializers.ModelSerializer):
-    name = serializers.CharField(
-        required=True,
-        help_text="Name of the rule",
-    )
-
-    action = serializers.JSONField(
-        required=True,
-        help_text="The action in the rule",
-    )
-
-    class Meta:
-        model = models.Rule
-        fields = [
-            "id",
-            "name",
-            "action",
-            "ruleset_id",
-        ]
-        read_only_fields = ["id", "created_at", "modified_at"]
-
-
-class RuleOutSerializer(serializers.Serializer):
-    id = serializers.IntegerField(
-        required=True,
-        help_text="ID of the ruleset",
-    )
-
-    name = serializers.CharField(
-        required=True,
-        help_text="Name of the rule",
-    )
-
-    action = serializers.JSONField(
-        default=dict,
-        help_text="The action in the rule",
-        allow_null=True,
-    )
-
-    fired_stats = serializers.ListField(
-        child=serializers.JSONField(),
-        required=True,
-        help_text="List of stats",
-    )
-
-    rulebook_id = serializers.PrimaryKeyRelatedField(
-        required=False,
-        allow_null=True,
-        queryset=models.Rulebook.objects.all(),
-        help_text="ID of the rulebook",
-    )
-
-    ruleset_id = serializers.PrimaryKeyRelatedField(
-        required=False,
-        allow_null=True,
-        queryset=models.Ruleset.objects.all(),
-        help_text="ID of the ruleset",
-    )
-
-    project_id = serializers.PrimaryKeyRelatedField(
-        required=False,
-        allow_null=True,
-        queryset=models.Project.objects.all(),
-        help_text="ID of the project",
-    )
 
 
 class AuditRuleSerializer(serializers.ModelSerializer):
@@ -335,6 +213,12 @@ class AuditActionSerializer(serializers.ModelSerializer):
         help_text="The fired timestamp of the action",
     )
 
+    status_message = serializers.CharField(
+        required=False,
+        allow_null=True,
+        help_text="Message of the action",
+    )
+
     class Meta:
         model = models.AuditAction
         fields = [
@@ -345,6 +229,7 @@ class AuditActionSerializer(serializers.ModelSerializer):
             "fired_at",
             "rule_fired_at",
             "audit_rule_id",
+            "status_message",
         ]
         read_only_fields = ["id"]
 
@@ -370,8 +255,9 @@ class AuditEventSerializer(serializers.ModelSerializer):
         help_text="The received timestamp of the event",
     )
 
-    payload = serializers.JSONField(
+    payload = YAMLSerializerField(
         required=False,
+        allow_null=True,
         help_text="The payload in the event",
     )
 
