@@ -11,7 +11,9 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from ansible_base.urls import urls as base_urls
+from ansible_base.resource_registry.urls import (
+    urlpatterns as resource_api_urls,
+)
 from django.urls import include, path
 from drf_spectacular.views import (
     SpectacularJSONAPIView,
@@ -20,6 +22,7 @@ from drf_spectacular.views import (
     SpectacularYAMLAPIView,
 )
 from rest_framework import routers
+from rest_framework_simplejwt import views as jwt_views
 
 from . import views
 
@@ -27,20 +30,20 @@ router = routers.SimpleRouter()
 router.register("extra-vars", views.ExtraVarViewSet)
 router.register("projects", views.ProjectViewSet)
 router.register("rulebooks", views.RulebookViewSet)
-router.register("rulesets", views.RulesetViewSet)
-router.register("rules", views.RuleViewSet)
 router.register("roles", views.RoleViewSet)
-router.register("tasks", views.TaskViewSet, basename="task")
 router.register("activations", views.ActivationViewSet)
 router.register("activation-instances", views.ActivationInstanceViewSet)
 router.register("audit-rules", views.AuditRuleViewSet)
 router.register("users", views.UserViewSet)
+router.register("event-streams", views.EventStreamViewSet)
 router.register(
     "users/me/awx-tokens",
     views.CurrentUserAwxTokenViewSet,
     basename="controller-token",
 )
 router.register("credentials", views.CredentialViewSet)
+router.register("credential-types", views.CredentialTypeViewSet)
+router.register("eda-credentials", views.EdaCredentialViewSet)
 router.register("decision-environments", views.DecisionEnvironmentViewSet)
 
 openapi_urls = [
@@ -67,13 +70,17 @@ openapi_urls = [
 ]
 
 v1_urls = [
+    path("", include(resource_api_urls)),
     path("", include(openapi_urls)),
     path("auth/session/login/", views.SessionLoginView.as_view()),
     path("auth/session/logout/", views.SessionLogoutView.as_view()),
+    path(
+        "auth/token/refresh/",
+        jwt_views.TokenRefreshView.as_view(),
+        name="token_refresh",
+    ),
     path("users/me/", views.CurrentUserView.as_view()),
     *router.urls,
-    # Experimental LDAP Integration https://issues.redhat.com/browse/AAP-16938
-    path("", include(base_urls)),
 ]
 
 urlpatterns = [
