@@ -12,10 +12,13 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 import logging
+import typing as tp
 
+import yaml
 from rest_framework import serializers
 
 from aap_eda.core import models
+from aap_eda.core.utils.credentials import validate_schema
 
 logger = logging.getLogger(__name__)
 
@@ -33,11 +36,11 @@ def check_if_rulebook_exists(rulebook_id: int) -> int:
 def check_if_de_exists(decision_environment_id: int) -> int:
     try:
         de = models.DecisionEnvironment.objects.get(pk=decision_environment_id)
-        if de.credential_id:
-            models.Credential.objects.get(pk=de.credential_id)
-    except models.Credential.DoesNotExist:
+        if de.eda_credential_id:
+            models.EdaCredential.objects.get(pk=de.eda_credential_id)
+    except models.EdaCredential.DoesNotExist:
         raise serializers.ValidationError(
-            f"Credential with id {de.credential_id} does not exist"
+            f"Credential with id {de.eda_credential_id} does not exist"
         )
     except models.DecisionEnvironment.DoesNotExist:
         raise serializers.ValidationError(
@@ -57,14 +60,71 @@ def check_if_extra_var_exists(extra_var_id: int) -> int:
     return extra_var_id
 
 
-def check_awx_tokens(user_id: int) -> int:
-    tokens = models.AwxToken.objects.filter(user_id=user_id).count()
-    if tokens == 0:
-        raise serializers.ValidationError("No controller token specified")
-    elif tokens > 1:
+def check_if_awx_token_exists(awx_token_id: int) -> int:
+    try:
+        models.AwxToken.objects.get(pk=awx_token_id)
+    except models.AwxToken.DoesNotExist:
         raise serializers.ValidationError(
-            "More than one controller token found, "
-            "currently only 1 token is supported"
+            f"AwxToken with id {awx_token_id} does not exist"
+        )
+    return awx_token_id
+
+
+def check_rulesets_require_token(
+    rulesets_data: list[dict[str, tp.Any]],
+) -> bool:
+    """Inspect rulesets data to determine if a token is required.
+
+    Return True if any of the rules has an action that requires a token.
+    """
+    required_actions = {"run_job_template", "run_workflow_template"}
+
+    for ruleset in rulesets_data:
+        for rule in ruleset.get("rules", []):
+            # When it is a single action dict
+            if any(
+                action_key in required_actions
+                for action_key in rule.get("action", {})
+            ):
+                return True
+
+            # When it is a list of actions
+            if any(
+                action_arg in required_actions
+                for action in rule.get("actions", [])
+                for action_arg in action
+            ):
+                return True
+
+    return False
+
+
+def is_extra_var_dict(extra_var: str):
+    try:
+        data = yaml.safe_load(extra_var)
+        if not isinstance(data, dict):
+            raise serializers.ValidationError(
+                "Extra var is not in object format"
+            )
+    except yaml.YAMLError:
+        raise serializers.ValidationError(
+            "Extra var must be in JSON or YAML format"
         )
 
-    return user_id
+
+def check_if_event_streams_exists(event_stream_ids: list[int]) -> list[int]:
+    for event_stream_id in event_stream_ids:
+        try:
+            models.EventStream.objects.get(pk=event_stream_id)
+        except models.EventStream.DoesNotExist:
+            raise serializers.ValidationError(
+                f"EventStream with id {event_stream_id} does not exist"
+            )
+    return event_stream_ids
+
+
+def check_if_schema_valid(schema: dict):
+    errors = validate_schema(schema)
+
+    if bool(errors):
+        raise serializers.ValidationError(errors)
