@@ -21,12 +21,12 @@ from .kubernetes import Engine as KubernetesEngine
 from .podman import Engine as PodmanEngine
 
 
-def new_container_engine(activation_id: str):
+def new_container_engine(activation_id: str, resource_prefix: str):
     """Activation service factory."""
     """Returns an activation object based on the deployment type"""
     # TODO: deployment type should not be plain strings
     if settings.DEPLOYMENT_TYPE == "podman":
-        return PodmanEngine(activation_id)
+        return PodmanEngine(activation_id, resource_prefix)
     if settings.DEPLOYMENT_TYPE == "k8s":
-        return KubernetesEngine(activation_id)
+        return KubernetesEngine(activation_id, resource_prefix)
     raise exceptions.InvalidDeploymentTypeError("Wrong deployment type")
