@@ -20,6 +20,7 @@ from rest_framework.test import APIClient
 
 from aap_eda.core import models
 from aap_eda.core.enums import Action, ResourceType
+from aap_eda.core.utils.credentials import inputs_to_display, inputs_to_store
 from tests.integration.constants import api_url_v1
 
 
@@ -27,15 +28,25 @@ from tests.integration.constants import api_url_v1
 # -------------------------------------
 @pytest.mark.django_db
 def test_list_projects(client: APIClient, check_permission_mock: mock.Mock):
+    cred_type = models.CredentialType.objects.create(
+        name="credtype",
+        inputs={},
+    )
     projects = models.Project.objects.bulk_create(
         [
             models.Project(
                 name="test-project-01",
                 url="https://git.example.com/acme/project-01",
                 git_hash="4673c67547cf6fe6a223a9dd49feb1d5f953449c",
-                credential=models.Credential.objects.create(
-                    name="credential1", username="me", secret="sec1"
+                eda_credential=models.EdaCredential.objects.create(
+                    name="credential1",
+                    credential_type=cred_type,
+                    inputs="sec1",
                 ),
+                signature_validation_credential=models.EdaCredential.objects.create(  # noqa: E501
+                    name="svcred1", credential_type=cred_type, inputs="sec2"
+                ),
+                scm_branch="main",
                 import_state=models.Project.ImportState.PENDING,
                 import_task_id="c8a7a0e3-05e7-4376-831a-6b8af80107bd",
             ),
@@ -44,6 +55,7 @@ def test_list_projects(client: APIClient, check_permission_mock: mock.Mock):
                 url="https://git.example.com/acme/project-02",
                 description="Project description.",
                 git_hash="06a71890b48189edc0b7afccf18285ec042ce302",
+                scm_refspec="refspec",
                 verify_ssl=False,
                 import_state=models.Project.ImportState.COMPLETED,
                 import_task_id="46e289a7-9dcc-4baa-a49a-a6ca756d9b71",
@@ -111,11 +123,21 @@ def test_list_projects_filter_name_none_exist(client: APIClient):
 
 @pytest.mark.django_db
 def test_retrieve_project(client: APIClient, check_permission_mock: mock.Mock):
+    cred_type = models.CredentialType.objects.create(
+        name="type1", inputs={"fields": [{"id": "user", "type": "string"}]}
+    )
+    cred_inputs = inputs_to_store({"user": "me"})
+    sv_cred_inputs = inputs_to_store({"user": "him"})
     project = models.Project.objects.create(
         name="test-project-01",
-        credential=models.Credential.objects.create(
-            name="credential1", username="me", secret="sec1"
+        eda_credential=models.EdaCredential.objects.create(
+            name="credential1", credential_type=cred_type, inputs=cred_inputs
         ),
+        signature_validation_credential=models.EdaCredential.objects.create(
+            name="svcred1", credential_type=cred_type, inputs=sv_cred_inputs
+        ),
+        scm_branch="main",
+        scm_refspec="ref1",
         url="https://git.example.com/acme/project-01",
         git_hash="4673c67547cf6fe6a223a9dd49feb1d5f953449c",
     )
@@ -143,7 +165,6 @@ def test_retrieve_project_failed_state(client: APIClient):
     data = response.json()
 
     assert data["import_state"] == "failed"
-    assert data["import_task_id"] == "3677eb4a-de4a-421a-a73b-411aa502484d"
     assert data["import_error"] == "Unexpected error. Please contact support."
 
     assert_project_data_details(data, project)
@@ -168,10 +189,26 @@ def test_create_project(
     job = mock.Mock(id=job_id)
     import_project_task.delay.return_value = job
 
+    cred_type = models.CredentialType.objects.create(
+        name="type1", inputs={"fields": [{"id": "user", "type": "string"}]}
+    )
+    cred_inputs = inputs_to_store({"user": "me"})
+    eda_credential = models.EdaCredential.objects.create(
+        name="credential1", credential_type=cred_type, inputs=cred_inputs
+    )
+    sv_cred_inputs = inputs_to_store({"user": "him"})
+    sv_credential = models.EdaCredential.objects.create(
+        name="svcred1", credential_type=cred_type, inputs=sv_cred_inputs
+    )
+
     bodies = [
         {
             "name": "test-project-01",
             "url": "https://git.example.com/acme/project-01",
+            "eda_credential_id": eda_credential.id,
+            "signature_validation_credential_id": sv_credential.id,
+            "scm_branch": "main",
+            "scm_refspec": "ref1",
         },
         {
             "name": "test-project-02",
@@ -353,14 +390,10 @@ def test_update_project_not_found(client: APIClient):
 
 @pytest.mark.django_db
 def test_update_project_with_400(client: APIClient):
-    credential = models.Credential.objects.create(
-        name="credential", username="me", secret="sec"
-    )
     first = models.Project.objects.create(
         name="test-project-01",
         url="https://git.example.com/acme/project-01",
         git_hash="4673c67547cf6fe6a223a9dd49feb1d5f953449c",
-        credential_id=credential.id,
     )
     second = models.Project.objects.create(
         name="test-project-02",
@@ -387,19 +420,27 @@ def test_update_project_with_400(client: APIClient):
     response = client.get(f"{api_url_v1}/projects/{first.id}/")
     data = {
         "name": "new project name",
-        "credential_id": 2,
+        "eda_credential_id": 2,
     }
     response = client.patch(f"{api_url_v1}/projects/{first.id}/", data=data)
     assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.data["errors"] == "Credential [2] not found"
+    assert response.data["errors"] == "EdaCredential [2] not found"
 
 
 @pytest.mark.django_db
 def test_partial_update_project(
     client: APIClient, check_permission_mock: mock.Mock
 ):
-    credential = models.Credential.objects.create(
-        name="credential1", username="me", secret="sec1"
+    cred_type = models.CredentialType.objects.create(
+        name="type1", inputs={"fields": [{"id": "user", "type": "string"}]}
+    )
+    cred_inputs = inputs_to_store({"user": "me"})
+    eda_credential = models.EdaCredential.objects.create(
+        name="credential1", credential_type=cred_type, inputs=cred_inputs
+    )
+    sv_cred_inputs = inputs_to_store({"user": "him"})
+    sv_credential = models.EdaCredential.objects.create(
+        name="svcred1", credential_type=cred_type, inputs=sv_cred_inputs
     )
 
     project = models.Project.objects.create(
@@ -407,14 +448,18 @@ def test_partial_update_project(
         url="https://git.example.com/acme/project-01",
         git_hash="4673c67547cf6fe6a223a9dd49feb1d5f953449c",
     )
-    assert project.credential_id is None
+    assert project.eda_credential_id is None
+    assert project.signature_validation_credential_id is None
     assert project.verify_ssl is True
 
     response = client.patch(
         f"{api_url_v1}/projects/{project.id}/",
         data={
             "name": "test-project-01-updated",
-            "credential_id": credential.id,
+            "eda_credential_id": eda_credential.id,
+            "signature_validation_credential_id": sv_credential.id,
+            "scm_branch": "main",
+            "scm_refspec": "ref1",
             "verify_ssl": False,
         },
     )
@@ -422,7 +467,8 @@ def test_partial_update_project(
 
     project = models.Project.objects.get(pk=project.id)
     assert project.name == "test-project-01-updated"
-    assert project.credential_id == credential.id
+    assert project.eda_credential_id == eda_credential.id
+    assert project.signature_validation_credential_id == sv_credential.id
     assert project.verify_ssl is False
 
     assert_project_data(response.json(), project)
@@ -462,51 +508,71 @@ DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
 
 
 def assert_project_data(data: Dict[str, Any], project: models.Project):
-    credential_id = project.credential.id if project.credential else None
-    verify_ssl = project.verify_ssl if project.verify_ssl is not None else True
-
-    project_to_data = model_to_data_common(project)
-    project_to_data["credential_id"] = credential_id
-    project_to_data["verify_ssl"] = verify_ssl
-
-    assert data == project_to_data
-
-
-def assert_project_data_details(data: Dict[str, Any], project: models.Project):
-    credential = project.credential
-    credential_data = (
-        {
-            "id": credential.id,
-            "name": credential.name,
-            "description": credential.description,
-            "credential_type": credential.credential_type,
-            "username": credential.username,
-        }
-        if credential
+    eda_credential_id = (
+        project.eda_credential.id if project.eda_credential else None
+    )
+    signature_validation_credential_id = (
+        project.signature_validation_credential.id
+        if project.signature_validation_credential
         else None
     )
     verify_ssl = project.verify_ssl if project.verify_ssl is not None else True
 
     project_to_data = model_to_data_common(project)
-    project_to_data["credential"] = credential_data
+    project_to_data["eda_credential_id"] = eda_credential_id
+    project_to_data[
+        "signature_validation_credential_id"
+    ] = signature_validation_credential_id
+    project_to_data["verify_ssl"] = verify_ssl
+
+    assert data == project_to_data
+
+
+def get_credential_details(credential: models.EdaCredential) -> dict:
+    credential.refresh_from_db()
+    return {
+        "id": credential.id,
+        "name": credential.name,
+        "description": credential.description,
+        "credential_type_id": credential.credential_type.id,
+        "managed": credential.managed,
+        "inputs": inputs_to_display(
+            credential.credential_type.inputs,
+            credential.inputs.get_secret_value(),
+        ),
+    }
+
+
+def assert_project_data_details(data: Dict[str, Any], project: models.Project):
+    eda_credential = project.eda_credential
+    eda_credential_data = (
+        get_credential_details(eda_credential) if eda_credential else None
+    )
+    sv_credential = project.signature_validation_credential
+    sv_credential_data = (
+        get_credential_details(sv_credential) if sv_credential else None
+    )
+    verify_ssl = project.verify_ssl if project.verify_ssl is not None else True
+
+    project_to_data = model_to_data_common(project)
+    project_to_data["eda_credential"] = eda_credential_data
+    project_to_data["signature_validation_credential"] = sv_credential_data
     project_to_data["verify_ssl"] = verify_ssl
 
     assert data == project_to_data
 
 
 def model_to_data_common(project: models.Project):
-    import_task_id = project.import_task_id
-    if import_task_id is not None:
-        import_task_id = str(import_task_id)
-
     return {
         "id": project.id,
         "url": project.url,
         "name": project.name,
         "description": project.description,
         "git_hash": project.git_hash,
+        "scm_type": project.scm_type,
+        "scm_branch": project.scm_branch,
+        "scm_refspec": project.scm_refspec,
         "import_state": project.import_state,
-        "import_task_id": import_task_id,
         "import_error": project.import_error,
         "created_at": project.created_at.strftime(DATETIME_FORMAT),
         "modified_at": project.modified_at.strftime(DATETIME_FORMAT),
