@@ -18,11 +18,36 @@ import pytest
 from django.conf import settings
 from pytest_redis import factories
 
+from aap_eda.core import models
 from aap_eda.core.tasking import Queue
 
 ADMIN_USERNAME = "test.admin"
 ADMIN_PASSWORD = "test.admin.123"
-
+INPUTS = {
+    "fields": [
+        {"id": "username", "label": "Username", "type": "string"},
+        {
+            "id": "password",
+            "label": "Password",
+            "type": "string",
+            "secret": True,
+        },
+        {
+            "id": "ssh_key_data",
+            "label": "SCM Private Key",
+            "type": "string",
+            "format": "ssh_private_key",
+            "secret": True,
+            "multiline": True,
+        },
+        {
+            "id": "ssh_key_unlock",
+            "label": "Private Key Passphrase",
+            "type": "string",
+            "secret": True,
+        },
+    ]
+}
 
 # fixture for pytest-redis plugin
 # like django-db for a running redis server
@@ -52,3 +77,14 @@ def caplog_factory(caplog):
         return caplog
 
     return _factory
+
+
+@pytest.fixture
+def credential_type() -> models.CredentialType:
+    """Return a default Credential Type."""
+    credential_type = models.CredentialType.objects.create(
+        name="default_credential_type", inputs=INPUTS, injectors={}
+    )
+    credential_type.refresh_from_db()
+
+    return credential_type
