@@ -12,13 +12,17 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
+import os
 from dataclasses import dataclass
+from pathlib import Path
 from unittest import mock
 
 import pytest
 from dateutil import parser
+from podman import PodmanClient
 from podman.errors import ContainerError, ImageNotFound
 from podman.errors.exceptions import APIError, NotFound
+from rq.timeouts import JobTimeoutException
 
 from aap_eda.core import models
 from aap_eda.core.enums import ActivationStatus
@@ -27,6 +31,7 @@ from aap_eda.services.activation.engine import messages
 from aap_eda.services.activation.engine.common import (
     AnsibleRulebookCmdLine,
     ContainerRequest,
+    Credential,
 )
 from aap_eda.services.activation.engine.exceptions import (
     ContainerCleanupError,
@@ -34,14 +39,17 @@ from aap_eda.services.activation.engine.exceptions import (
     ContainerImagePullError,
     ContainerNotFoundError,
     ContainerStartError,
+    ContainerUpdateLogsError,
 )
-from aap_eda.services.activation.engine.podman import Engine
+from aap_eda.services.activation.engine.podman import Engine, get_podman_client
+
+DATA_DIR = Path(__file__).parent / "data"
 
 
 @dataclass
 class InitData:
     activation: models.Activation
-    activation_instance: models.ActivationInstance
+    activation_instance: models.RulebookProcess
 
 
 @pytest.fixture()
@@ -57,7 +65,7 @@ def init_data():
         name="activation",
         user=user,
     )
-    activation_instance = models.ActivationInstance.objects.create(
+    activation_instance = models.RulebookProcess.objects.create(
         name="test-instance",
         log_read_at=parser.parse("2023-10-30T19:18:48.362883381Z"),
         activation=activation,
@@ -73,6 +81,9 @@ def get_ansible_rulebook_cmdline(data: InitData):
     return AnsibleRulebookCmdLine(
         ws_url="ws://localhost:8000/api/eda/ws/ansible-rulebook",
         ws_ssl_verify="no",
+        ws_token_url="http://localhost:8000/api/eda/v1/auth/token/refresh",
+        ws_access_token="access",
+        ws_refresh_token="refresh",
         id=data.activation.id,
         log_level="-v",
         heartbeat=5,
@@ -83,9 +94,36 @@ def get_request(data: InitData):
     return ContainerRequest(
         name="test-request",
         image_url="quay.io/ansible/ansible-rulebook:main",
-        activation_instance_id=data.activation_instance.id,
-        activation_id=data.activation.id,
+        rulebook_process_id=data.activation_instance.id,
+        process_parent_id=data.activation.id,
         cmdline=get_ansible_rulebook_cmdline(data),
+        ports=[("localhost", 8080)],
+        mem_limit="8G",
+        mounts=[{"/dev": "/opt"}],
+        env_vars={"a": 1},
+        extra_args={"b": 2},
+    )
+
+
+def get_request_with_never_pull_policy(data: InitData):
+    return ContainerRequest(
+        name="test-request",
+        image_url="quay.io/ansible/ansible-rulebook:main",
+        rulebook_process_id=data.activation_instance.id,
+        process_parent_id=data.activation.id,
+        cmdline=get_ansible_rulebook_cmdline(data),
+        pull_policy="Never",
+    )
+
+
+def get_request_with_credential(data: InitData):
+    return ContainerRequest(
+        name="test-request",
+        image_url="quay.io/ansible/ansible-rulebook:main",
+        rulebook_process_id=data.activation_instance.id,
+        process_parent_id=data.activation.id,
+        cmdline=get_ansible_rulebook_cmdline(data),
+        credential=Credential(username="me", secret="secret"),
     )
 
 
@@ -106,6 +144,39 @@ def podman_engine(init_data):
         )
 
         yield engine
+
+
+def test_get_podman_client(settings):
+    settings.PODMAN_SOCKET_URL = None
+    uid_0_mock = mock.Mock(return_value=0)
+
+    with mock.patch("os.getuid", uid_0_mock):
+        client = get_podman_client()
+        assert client.api.base_url.netloc == "%2Frun%2Fpodman%2Fpodman.sock"
+
+    client = get_podman_client()
+    xdg_runtime_dir = os.getenv(
+        "XDG_RUNTIME_DIR", f"%2Frun%2Fuser%2F{os.getuid()}"
+    )
+    # Replace any '/'s from XDG_RUNTIME_DIR.
+    xdg_runtime_dir = xdg_runtime_dir.replace("/", "%2F")
+
+    assert (
+        client.api.base_url.netloc
+        == f"{xdg_runtime_dir}%2Fpodman%2Fpodman.sock"
+    )
+
+
+def test_get_podman_client_with_exception(settings):
+    def raise_error(*args, **kwargs):
+        raise ValueError("Failed to initialize client")
+
+    with mock.patch.object(PodmanClient, "__init__", side_effect=raise_error):
+        with pytest.raises(
+            ContainerEngineInitError,
+            match="Failed to initialize client",
+        ):
+            get_podman_client()
 
 
 @pytest.mark.django_db
@@ -135,10 +206,8 @@ def test_engine_start(init_data, podman_engine):
     engine.start(request, log_handler)
 
     engine.client.containers.run.assert_called_once()
-    assert models.ActivationInstanceLog.objects.count() == 4
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
-        "is started."
-    )
+    assert models.RulebookProcessLog.objects.count() == 4
+    assert models.RulebookProcessLog.objects.last().log.endswith("is started.")
 
 
 @pytest.mark.django_db
@@ -180,6 +249,11 @@ def test_engine_start_with_credential(init_data, podman_engine):
         remove=True,
         detach=True,
         name=request.name,
+        ports={"8080/tcp": 8080},
+        mem_limit="8G",
+        mounts=request.mounts,
+        environment={"a": 1},
+        b=2,
     )
 
 
@@ -212,19 +286,60 @@ def test_engine_start_with_image_not_found_exception(init_data, podman_engine):
     request = get_request(init_data)
     request.credential = credential
 
-    def raise_error(*args, **kwargs):
+    def raise_not_found_error(*args, **kwargs):
         raise ImageNotFound("Image not found")
 
-    engine.client.images.pull.side_effect = raise_error
+    engine.client.images.pull.side_effect = raise_not_found_error
 
     with pytest.raises(
         ContainerImagePullError, match=f"Image {request.image_url} not found"
     ):
         engine.start(request, log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         f"Image {request.image_url} not found"
     )
+
+
+@pytest.mark.django_db
+def test_engine_start_with_image_api_exception(init_data, podman_engine):
+    credential = models.Credential.objects.create(
+        name="credential1", username="me", secret="sec1"
+    )
+    engine = podman_engine
+    log_handler = DBLogger(init_data.activation_instance.id)
+    request = get_request(init_data)
+    request.credential = credential
+
+    def raise_api_error(*args, **kwargs):
+        raise APIError("Image not found")
+
+    engine.client.images.pull.side_effect = raise_api_error
+
+    with pytest.raises(ContainerStartError, match="Image not found"):
+        engine.start(request, log_handler)
+
+
+@pytest.mark.django_db
+def test_engine_start_with_image_pull_timeout_exception(
+    init_data, podman_engine
+):
+    credential = models.Credential.objects.create(
+        name="credential1", username="me", secret="sec1"
+    )
+    engine = podman_engine
+    log_handler = DBLogger(init_data.activation_instance.id)
+    request = get_request(init_data)
+    request.credential = credential
+    error = "Task exceeded maximum timeout value 120 seconds"
+
+    def raise_timeout_error(*args, **kwargs):
+        raise JobTimeoutException(error)
+
+    engine.client.images.pull.side_effect = raise_timeout_error
+
+    with pytest.raises(ContainerImagePullError, match=error):
+        engine.start(request, log_handler)
 
 
 @pytest.mark.django_db
@@ -248,7 +363,7 @@ def test_engine_start_with_image_pull_exception(init_data, podman_engine):
         with pytest.raises(ContainerImagePullError, match=msg):
             engine.start(request, log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(msg)
+    assert models.RulebookProcessLog.objects.last().log.endswith(msg)
 
 
 @pytest.mark.django_db
@@ -276,8 +391,19 @@ def test_engine_start_with_containers_run_exception(init_data, podman_engine):
 
     assert (
         "Container Start Error:"
-        in models.ActivationInstanceLog.objects.last().log
+        in models.RulebookProcessLog.objects.last().log
     )
+
+
+@pytest.mark.django_db
+def test_engine_start_with_never_pull_policy_request(init_data, podman_engine):
+    engine = podman_engine
+    request = get_request_with_never_pull_policy(init_data)
+    log_handler = DBLogger(init_data.activation_instance.id)
+
+    engine.start(request, log_handler)
+
+    engine.client.containers.run.assert_called_once()
 
 
 @pytest.mark.django_db
@@ -296,6 +422,26 @@ def test_engine_get_status(podman_engine):
         pod_id="container_id"
     )
 
+    # when status in "created"
+    container_mock.status = "created"
+    container_mock.attrs = {"State": {"Error": None}}
+    activation_status = engine.get_status("container_id")
+
+    assert activation_status.status == ActivationStatus.FAILED
+
+    for unexpected_state in [
+        "paused",
+        "restarting",
+        "removing",
+        "dead",
+        "configured",
+        "unknown",
+    ]:
+        container_mock.status = unexpected_state
+        activation_status = engine.get_status("container_id")
+
+        assert activation_status.status == ActivationStatus.FAILED
+
     # when status in "exited"
     container_mock.status = "exited"
     expects = [
@@ -304,7 +450,15 @@ def test_engine_get_status(podman_engine):
     ]
 
     for key, value in expects:
-        container_mock.attrs.get.return_value.get.return_value = key
+        container_mock.attrs = {"State": {"ExitCode": key}}
+        activation_status = engine.get_status("container_id")
+
+        assert activation_status.status == value
+
+    # when status is stopped
+    container_mock.status = "stopped"
+    for key, value in expects:
+        container_mock.attrs = {"State": {"ExitCode": key}}
         activation_status = engine.get_status("container_id")
 
         assert activation_status.status == value
@@ -329,7 +483,7 @@ def test_engine_cleanup(init_data, podman_engine):
 
     engine.cleanup("100", log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         "Container 100 is cleaned up."
     )
 
@@ -348,13 +502,32 @@ def test_engine_cleanup_with_not_found_exception(init_data, podman_engine):
 
     engine.cleanup("100", log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         "Container 100 not found."
     )
 
 
 @pytest.mark.django_db
-def test_engine_cleanup_with_api_exception(init_data, podman_engine):
+def test_engine_cleanup_with_remove_exception(init_data, podman_engine):
+    engine = podman_engine
+    log_handler = DBLogger(init_data.activation_instance.id)
+
+    err_msg = "Not found"
+
+    def raise_error(*args, **kwargs):
+        raise APIError(err_msg)
+
+    container_mock = mock.Mock()
+    engine.client.containers.get.return_value = container_mock
+    container_mock.logs.return_value = []
+    container_mock.remove.side_effect = raise_error
+
+    with pytest.raises(ContainerCleanupError, match=err_msg):
+        engine.cleanup("100", log_handler)
+
+
+@pytest.mark.django_db
+def test_engine_cleanup_with_get_exception(init_data, podman_engine):
     engine = podman_engine
     log_handler = DBLogger(init_data.activation_instance.id)
 
@@ -403,10 +576,10 @@ def test_engine_update_logs(init_data, podman_engine):
 
     engine.update_logs("100", log_handler)
 
-    assert models.ActivationInstanceLog.objects.count() == len(
+    assert models.RulebookProcessLog.objects.count() == len(
         container_mock.logs.return_value
     )
-    assert models.ActivationInstanceLog.objects.last().log == f"{message}"
+    assert models.RulebookProcessLog.objects.last().log == f"{message}"
 
     init_data.activation_instance.refresh_from_db()
     assert init_data.activation_instance.log_read_at > init_log_read_at
@@ -420,6 +593,20 @@ def test_engine_update_logs_with_container_not_found(init_data, podman_engine):
     engine.client.containers.exists.return_value = None
     engine.update_logs("100", log_handler)
 
-    assert models.ActivationInstanceLog.objects.last().log.endswith(
+    assert models.RulebookProcessLog.objects.last().log.endswith(
         "Container 100 not found."
     )
+
+
+@pytest.mark.django_db
+def test_engine_update_logs_with_exception(init_data, podman_engine):
+    engine = podman_engine
+    log_handler = DBLogger(init_data.activation_instance.id)
+
+    def raise_error(*args, **kwargs):
+        raise APIError("Not found")
+
+    engine.client.containers.exists.side_effect = raise_error
+
+    with pytest.raises(ContainerUpdateLogsError, match="Not found"):
+        engine.update_logs("100", log_handler)
