@@ -21,15 +21,15 @@ from _pytest.logging import LogCaptureFixture
 from pytest_django.fixtures import SettingsWrapper
 from pytest_lazyfixture import lazy_fixture
 
-from aap_eda.core import models
-from aap_eda.core.enums import ActivationStatus
+from aap_eda.core import enums, models
 from aap_eda.services.activation.engine import exceptions as engine_exceptions
-from aap_eda.services.activation.engine.common import ContainerEngine
+from aap_eda.services.activation.engine.common import (
+    ContainerEngine,
+    ContainerRequest,
+)
 from aap_eda.services.activation.manager import (
-    ACTIVATION_PATH,
     LOGGER,
     ActivationManager,
-    AnsibleRulebookCmdLine,
     exceptions,
 )
 
@@ -41,7 +41,7 @@ def apply_settings(settings: SettingsWrapper, **kwargs):
 
 
 @pytest.fixture
-def default_rulebook() -> models.Rulebook:
+def rulebook_with_job_template() -> models.Rulebook:
     """Return a default rulebook."""
     rulesets = """
 ---
@@ -54,7 +54,7 @@ def default_rulebook() -> models.Rulebook:
     - name: Say Hello
       condition: event.i == 1
       action:
-        debug:
+        run_job_template:
           msg: "Hello World!"
 
 """
@@ -71,36 +71,11 @@ def eda_caplog(caplog_factory) -> LogCaptureFixture:
 
 
 @pytest.fixture
-def default_user() -> models.User:
-    """Return a default user."""
-    user = models.User.objects.create(
-        username="test.user",
-        password="test.user.123",
-        email="test.user@localhost",
-    )
-
-    models.AwxToken.objects.create(
-        name="test-token", user=user, token="supersecret"
-    )
-
-    return user
-
-
-@pytest.fixture
-def decision_environment() -> models.DecisionEnvironment:
-    """Return a default decision environment."""
-    return models.DecisionEnvironment.objects.create(
-        name="test-decision-environment",
-        image_url="localhost:14000/test-image-url",
-    )
-
-
-@pytest.fixture
 def activation_with_instance(
     basic_activation: models.Activation,
 ) -> models.Activation:
     """Return an activation with an instance."""
-    models.ActivationInstance.objects.create(activation=basic_activation)
+    models.RulebookProcess.objects.create(activation=basic_activation)
     return basic_activation
 
 
@@ -108,9 +83,9 @@ def activation_with_instance(
 def running_activation(activation_with_instance: models.Activation):
     """Return a running activation."""
     activation = activation_with_instance
-    activation.status = ActivationStatus.RUNNING
+    activation.status = enums.ActivationStatus.RUNNING
     activation.save(update_fields=["status"])
-    activation.latest_instance.status = ActivationStatus.RUNNING
+    activation.latest_instance.status = enums.ActivationStatus.RUNNING
     activation.latest_instance.activation_pod_id = "test-pod-id"
     activation.latest_instance.save(
         update_fields=["status", "activation_pod_id"],
@@ -121,18 +96,41 @@ def running_activation(activation_with_instance: models.Activation):
 @pytest.fixture
 def basic_activation(
     default_user: models.User,
-    decision_environment: models.DecisionEnvironment,
+    default_decision_environment: models.DecisionEnvironment,
     default_rulebook: models.Rulebook,
 ) -> models.Activation:
     """Return the minimal activation."""
     return models.Activation.objects.create(
         name="test-activation",
         user=default_user,
-        decision_environment=decision_environment,
+        decision_environment=default_decision_environment,
+        rulebook=default_rulebook,
+        # rulebook_rulesets is populated by the serializer
+        rulebook_rulesets=default_rulebook.rulesets,
+        log_level=enums.RulebookProcessLogLevel.INFO,
+    )
+
+
+@pytest.fixture
+def new_activation_with_instance(
+    default_user: models.User,
+    default_decision_environment: models.DecisionEnvironment,
+    default_rulebook: models.Rulebook,
+) -> models.Activation:
+    """Return an activation with an instance."""
+    activation = models.Activation.objects.create(
+        name="new_activation_with_instance",
+        user=default_user,
+        decision_environment=default_decision_environment,
         rulebook=default_rulebook,
         # rulebook_rulesets is populated by the serializer
         rulebook_rulesets=default_rulebook.rulesets,
     )
+    models.RulebookProcess.objects.create(
+        activation=activation,
+        status=enums.ActivationStatus.RUNNING,
+    )
+    return activation
 
 
 @pytest.fixture
@@ -141,26 +139,23 @@ def container_engine_mock() -> MagicMock:
 
 
 @pytest.mark.django_db
-def test_build_cmdline(
+def test_get_container_request(
     activation_with_instance: models.Activation,
     settings: SettingsWrapper,
 ):
     """Test build_cmdline."""
     override_settings = {
         "WEBSOCKET_BASE_URL": "ws://localhost:8000",
-        "ANSIBLE_RULEBOOK_LOG_LEVEL": "-vv",
         "WEBSOCKET_SSL_VERIFY": "no",
         "RULEBOOK_LIVENESS_CHECK_SECONDS": 73,
     }
     apply_settings(settings, **override_settings)
     activation_manager = ActivationManager(activation_with_instance)
-    cmdline = activation_manager._build_cmdline()
-    assert isinstance(cmdline, AnsibleRulebookCmdLine)
-    assert (
-        cmdline.ws_url
-        == override_settings["WEBSOCKET_BASE_URL"] + ACTIVATION_PATH
-    )
-    assert cmdline.log_level == override_settings["ANSIBLE_RULEBOOK_LOG_LEVEL"]
+    request = activation_manager._get_container_request()
+    assert isinstance(request, ContainerRequest)
+    cmdline = request.cmdline
+    assert cmdline.ws_url.startswith(override_settings["WEBSOCKET_BASE_URL"])
+    assert cmdline.log_level == "-v"
     assert cmdline.ws_ssl_verify == override_settings["WEBSOCKET_SSL_VERIFY"]
     assert (
         cmdline.heartbeat
@@ -170,18 +165,11 @@ def test_build_cmdline(
 
 
 @pytest.mark.django_db
-def test_build_cmdline_no_instance(basic_activation):
+def test_get_container_request_no_instance(basic_activation):
     """Test build_cmdline when no instance exists."""
     activation_manager = ActivationManager(basic_activation)
     with pytest.raises(exceptions.ActivationManagerError):
-        activation_manager._build_cmdline()
-
-
-@pytest.mark.django_db
-def test_build_credential_inexistent(basic_activation):
-    """Test build_credential when no credential exists."""
-    activation_manager = ActivationManager(basic_activation)
-    assert activation_manager._build_credential() is None
+        activation_manager._get_container_request()
 
 
 @pytest.mark.django_db
@@ -208,14 +196,16 @@ def test_start_disabled_activation(activation_with_instance, eda_caplog):
 
 
 @pytest.mark.django_db
-def test_start_no_awx_token(basic_activation):
+def test_start_no_awx_token(basic_activation, rulebook_with_job_template):
     """Test start verb when no AWX token is present."""
+    basic_activation.rulebook = rulebook_with_job_template
+    basic_activation.save(update_fields=["rulebook"])
     activation_manager = ActivationManager(basic_activation)
     basic_activation.user.awxtoken_set.all().delete()
     with pytest.raises(exceptions.ActivationStartError) as exc:
         activation_manager.start()
-    assert basic_activation.status == ActivationStatus.ERROR
-    assert "No controller token specified" in str(exc.value)
+    assert basic_activation.status == enums.ActivationStatus.ERROR
+    assert "The rulebook requires an Awx Token." in str(exc.value)
     assert str(exc.value) in basic_activation.status_message
 
 
@@ -226,7 +216,7 @@ def test_start_no_decision_environment(basic_activation):
     basic_activation.decision_environment.delete()
     with pytest.raises(exceptions.ActivationStartError) as exc:
         activation_manager.start()
-    assert basic_activation.status == ActivationStatus.ERROR
+    assert basic_activation.status == enums.ActivationStatus.ERROR
     assert "decision_environment" in str(exc.value)
     assert "This field may not be null" in str(exc.value)
     assert str(exc.value) in basic_activation.status_message
@@ -244,13 +234,13 @@ def test_start_already_running(
         container_engine=container_engine_mock,
     )
     container_engine_mock.get_status.return_value = MagicMock(
-        status=ActivationStatus.RUNNING,
+        status=enums.ActivationStatus.RUNNING,
     )
 
     activation_manager.start()
     assert container_engine_mock.get_status.called
     assert "already running" in eda_caplog.text
-    assert running_activation.status == ActivationStatus.RUNNING
+    assert running_activation.status == enums.ActivationStatus.RUNNING
 
 
 @pytest.mark.django_db
@@ -269,8 +259,11 @@ def test_start_first_run(
     assert container_engine_mock.start.called
     assert container_engine_mock.update_logs.called
     assert "Starting" in eda_caplog.text
-    assert basic_activation.status == ActivationStatus.RUNNING
-    assert basic_activation.latest_instance.status == ActivationStatus.RUNNING
+    assert basic_activation.status == enums.ActivationStatus.RUNNING
+    assert (
+        basic_activation.latest_instance.status
+        == enums.ActivationStatus.RUNNING
+    )
     assert basic_activation.latest_instance.activation_pod_id == "test-pod-id"
     assert basic_activation.restart_count == 0
 
@@ -291,9 +284,10 @@ def test_start_restart(
     assert container_engine_mock.start.called
     assert container_engine_mock.update_logs.called
     assert "Starting" in eda_caplog.text
-    assert running_activation.status == ActivationStatus.RUNNING
+    assert running_activation.status == enums.ActivationStatus.RUNNING
     assert (
-        running_activation.latest_instance.status == ActivationStatus.RUNNING
+        running_activation.latest_instance.status
+        == enums.ActivationStatus.RUNNING
     )
     assert (
         running_activation.latest_instance.activation_pod_id == "test-pod-id"
@@ -332,8 +326,10 @@ def test_stop_already_stopped(
         db_instance=activation_with_instance,
     )
 
-    activation_with_instance.status = ActivationStatus.STOPPED
-    activation_with_instance.latest_instance.status = ActivationStatus.STOPPED
+    activation_with_instance.status = enums.ActivationStatus.STOPPED
+    activation_with_instance.latest_instance.status = (
+        enums.ActivationStatus.STOPPED
+    )
     activation_with_instance.latest_instance.save(
         update_fields=["status"],
     )
@@ -361,9 +357,10 @@ def test_stop_running(
     assert "Stopping" in eda_caplog.text
     assert "Cleanup operation requested" in eda_caplog.text
     assert "Activation stopped." in eda_caplog.text
-    assert running_activation.status == ActivationStatus.STOPPED
+    assert running_activation.status == enums.ActivationStatus.STOPPED
     assert (
-        running_activation.latest_instance.status == ActivationStatus.STOPPED
+        running_activation.latest_instance.status
+        == enums.ActivationStatus.STOPPED
     )
     assert running_activation.latest_instance.activation_pod_id is None
     assert running_activation.status_message == "Stop requested by user."
@@ -390,9 +387,10 @@ def test_stop_no_pod(
     assert "Stopping" in eda_caplog.text
     assert "Cleanup operation requested" in eda_caplog.text
     assert "Activation stopped." in eda_caplog.text
-    assert running_activation.status == ActivationStatus.STOPPED
+    assert running_activation.status == enums.ActivationStatus.STOPPED
     assert (
-        running_activation.latest_instance.status == ActivationStatus.STOPPED
+        running_activation.latest_instance.status
+        == enums.ActivationStatus.STOPPED
     )
     assert running_activation.latest_instance.activation_pod_id is None
     assert running_activation.status_message == "Stop requested by user."
@@ -416,7 +414,7 @@ def test_stop_pending(
     assert not container_engine_mock.get_status.called
     assert not container_engine_mock.cleanup.called
     assert "No instance found" in eda_caplog.text
-    assert basic_activation.status == ActivationStatus.STOPPED
+    assert basic_activation.status == enums.ActivationStatus.STOPPED
     assert basic_activation.latest_instance is None
 
 
@@ -432,16 +430,17 @@ def test_stop_stopped_instance_running(
         db_instance=running_activation,
     )
 
-    running_activation.status = ActivationStatus.STOPPED
+    running_activation.status = enums.ActivationStatus.STOPPED
     running_activation.save(update_fields=["status"])
 
     activation_manager.stop()
     assert "Stopping" in eda_caplog.text
     assert container_engine_mock.get_status.called
     assert container_engine_mock.cleanup.called
-    assert running_activation.status == ActivationStatus.STOPPED
+    assert running_activation.status == enums.ActivationStatus.STOPPED
     assert (
-        running_activation.latest_instance.status == ActivationStatus.STOPPED
+        running_activation.latest_instance.status
+        == enums.ActivationStatus.STOPPED
     )
     assert running_activation.latest_instance.activation_pod_id is None
 
@@ -458,22 +457,23 @@ def test_stop_stopped_pod_running(
         db_instance=running_activation,
     )
 
-    running_activation.status = ActivationStatus.STOPPED
+    running_activation.status = enums.ActivationStatus.STOPPED
     running_activation.save(update_fields=["status"])
-    running_activation.latest_instance.status = ActivationStatus.STOPPED
+    running_activation.latest_instance.status = enums.ActivationStatus.STOPPED
     running_activation.latest_instance.save(update_fields=["status"])
 
     container_engine_mock.get_status.return_value = MagicMock(
-        status=ActivationStatus.RUNNING,
+        status=enums.ActivationStatus.RUNNING,
     )
 
     activation_manager.stop()
     assert "Stopping" in eda_caplog.text
     assert container_engine_mock.get_status.called
     assert container_engine_mock.cleanup.called
-    assert running_activation.status == ActivationStatus.STOPPED
+    assert running_activation.status == enums.ActivationStatus.STOPPED
     assert (
-        running_activation.latest_instance.status == ActivationStatus.STOPPED
+        running_activation.latest_instance.status
+        == enums.ActivationStatus.STOPPED
     )
     assert running_activation.latest_instance.activation_pod_id is None
 
@@ -501,22 +501,22 @@ def test_delete_already_deleted(
     [
         pytest.param(
             lazy_fixture("activation_with_instance"),
-            ActivationStatus.STOPPED,
+            enums.ActivationStatus.STOPPED,
             id="stopped",
         ),
         pytest.param(
             lazy_fixture("activation_with_instance"),
-            ActivationStatus.ERROR,
+            enums.ActivationStatus.ERROR,
             id="error",
         ),
         pytest.param(
             lazy_fixture("activation_with_instance"),
-            ActivationStatus.FAILED,
+            enums.ActivationStatus.FAILED,
             id="failed",
         ),
         pytest.param(
             lazy_fixture("activation_with_instance"),
-            ActivationStatus.COMPLETED,
+            enums.ActivationStatus.COMPLETED,
             id="completed",
         ),
     ],
@@ -524,7 +524,7 @@ def test_delete_already_deleted(
 @pytest.mark.django_db
 def test_delete_no_pod(
     activation: models.Activation,
-    status: ActivationStatus,
+    status: enums.ActivationStatus,
     container_engine_mock: MagicMock,
     eda_caplog: LogCaptureFixture,
 ):
@@ -585,3 +585,19 @@ def test_delete_with_exception(
     assert (
         models.Activation.objects.filter(id=running_activation.id).count() == 0
     )
+
+
+@pytest.mark.django_db
+def test_start_max_running_activations(
+    new_activation_with_instance: models.Activation,
+    basic_activation: models.Activation,
+    settings: SettingsWrapper,
+    eda_caplog: LogCaptureFixture,
+):
+    """Test start verb when max running activations is reached."""
+    apply_settings(settings, MAX_RUNNING_ACTIVATIONS=1)
+    activation_manager = ActivationManager(basic_activation)
+
+    with pytest.raises(exceptions.MaxRunningProcessesError):
+        activation_manager.start()
+    assert "No capacity to start a new rulebook process" in eda_caplog.text
